Initial commit
The code was downloaded today as a zip folder from:

https://plugins.jetbrains.com/plugin/1800-database-navigator

The build seems to be from Feb 16 2018

I believe they don't provide it as a source code repo (single dev). It
was probably hosted on google code and there are a few old forks of it
on repo from when google code closed.

According to the jetbrains plugin page, the code is licensed under Apache 2.0.